refactor: 나의 팟 페이지 상태 바 수정
